+++ packages/components/src/layouts/Navigation/utils/StatusBarUtils.native.ts
@@ -19,4 +19,3 @@ export const setHidden = (
 ): void => {
   StatusBar.setHidden(hidden, animation);
 };
-
