+++ packages/kit/src/provider/Container/KeyboardContainer/index.ios.tsx
@@ -1,0 +1,3 @@
+export function KeyboardContainer() {
+  return null;
+}
